requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.9
django-allauth==0.54.0
django-cors-headers==3.13.0
django-environ==0.12.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.10
google-auth==2.40.3     
gunicorn==23.0.0          
idna==3.10
oauthlib==3.3.1
packaging==25.0
psycopg2-binary==2.9.10   
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
dj-database-url==0.5.0
whitenoise
